requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==23.2.0
azure-cognitiveservices-speech==1.38.0
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-speech==2.26.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
gTTS==2.5.2
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
keras==3.4.1
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
playsound==1.3.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyglet==2.0.16
Pygments==2.18.0
python-dotenv==1.0.1
pyttsx3==2.90
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
srsly==2.4.8
srt==3.5.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
